@@ -4,18 +4,18 @@ version: 5.0.3
 homepage: https://github.com/ShadyBoukhary/flutter_clean_architecture
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.2.3 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  logging: ^1.0.0
-  meta: ^1.3.0
-  provider: ^6.0.1
-  responsive_builder: ^0.4.1
-  rxdart: ^0.26.0
+  logging: ^1.2.0
+  meta: ^1.10.0
+  provider: ^6.1.1
+  responsive_builder: ^0.7.0
+  rxdart: ^0.27.7
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.1
