@@ -2,15 +2,15 @@ import '../../../domain/usecases/get_user_usecase.dart';
 import 'package:flutter_clean_architecture/flutter_clean_architecture.dart';
 
 class HomePresenter extends Presenter {
-  Function getUserOnNext;
-  Function getUserOnComplete;
-  Function getUserOnError;
+  Function? getUserOnNext;
+  Function? getUserOnComplete;
+  Function? getUserOnError;
 
   final GetUserUseCase getUserUseCase;
   HomePresenter(usersRepo) : getUserUseCase = GetUserUseCase(usersRepo);
 
   void getUser(String uid) {
-    // execute getUseruserCase
+    // execute getUserUseCase
     getUserUseCase.execute(
         _GetUserUseCaseObserver(this), GetUserUseCaseParams(uid));
   }
@@ -26,19 +26,16 @@ class _GetUserUseCaseObserver extends Observer<GetUserUseCaseResponse> {
   _GetUserUseCaseObserver(this.presenter);
   @override
   void onComplete() {
-    assert(presenter.getUserOnComplete != null);
-    presenter.getUserOnComplete();
+    presenter.getUserOnComplete?.call();
   }
 
   @override
   void onError(e) {
-    assert(presenter.getUserOnError != null);
-    presenter.getUserOnError(e);
+    presenter.getUserOnError?.call(e);
   }
 
   @override
   void onNext(response) {
-    assert(presenter.getUserOnNext != null);
-    presenter.getUserOnNext(response.user);
+    presenter.getUserOnNext?.call(response?.user);
   }
 }
