@@ -2,12 +2,12 @@ import '../../domain/entities/user.dart';
 import '../../domain/repositories/users_repository.dart';
 
 class DataUsersRepository extends UsersRepository {
-  List<User> users;
-  // sigleton
+  List<User>? users;
+  // singleton
   static final DataUsersRepository _instance = DataUsersRepository._internal();
   DataUsersRepository._internal() {
     users = <User>[];
-    users.addAll([
+    users?.addAll([
       User('test-uid', 'John Smith', 18),
       User('test-uid2', 'John Doe', 22)
     ]);
@@ -17,13 +17,17 @@ class DataUsersRepository extends UsersRepository {
   @override
   Future<List<User>> getAllUsers() async {
     // Here, do some heavy work lke http requests, async tasks, etc to get data
-    return users;
+    return users ?? [];
   }
 
   @override
   Future<User> getUser(String uid) async {
     // Here, do some heavy work lke http requests, async tasks, etc to get data
 
-    return users.firstWhere((user) => user.uid == uid);
+    return users?.firstWhere((user) => user.uid == uid) ?? User(
+      'test-uid',
+      'John Smith',
+      18,
+    );
   }
 }
